Refactor dock layout module
