+++ src/plugin/amulet/editor/dock/_layout.py
@@ -1,12 +1,45 @@
-"""A module to manage everything about layouts.
-Registering layouts, adding layout buttons and enabling layouts"""
+"""
+Classes and functions to manage the docking system.
+"""
 
 from __future__ import annotations
-from threading import Lock
 from dataclasses import dataclass
+from threading import Lock
 import re
 
-from .layout import LayoutConfig
+from PySide6.QtCore import Qt, QPoint, QSize
+
+
+@dataclass(frozen=True)
+class WidgetConfig:
+    identifier: str
+
+
+@dataclass(frozen=True)
+class WidgetStackConfig:
+    widgets: tuple[WidgetConfig, ...]
+
+
+@dataclass(frozen=True)
+class SplitterConfig:
+    first: SplitterConfig | WidgetStackConfig
+    second: SplitterConfig | WidgetStackConfig
+    orientation: Qt.Orientation
+    weight: float
+
+
+@dataclass(frozen=True)
+class WindowConfig:
+    origin: QPoint | None
+    size: QSize | None
+    layout: SplitterConfig | WidgetStackConfig
+
+
+@dataclass(frozen=True)
+class LayoutConfig:
+    main_window: WindowConfig
+    sub_windows: tuple[WindowConfig, ...]
+
 
 # my_namespace.my_layout
 # my_namespace.my_group.my_layout
